Begin parsing text inputs for later getting photos from API
Put the first of the methods in place for updating the STORE on change
of state. findLetterCounts method is fuctioning perfectly. All letters,
numerals, and the six supported punctuation marks are counted and a
calculation is made to determine how many (max) letter photos to fetch
from the API. The first time a letter is found in the text, 25 is added
to the photo count to fetch. For each subsequent match, 10 is added to
it. So characters found more often will potentially get more pictures to
choose from, while those which don't occur wil not have photos fetched.

All of this works perfectly when the user clicks Back and edits the text
in any way.