All API calls complete and data stowed correctly
For all characters used in the text inputs, the API calls now fetch as
many photo URLs as are available, up to the limits set in the app. They
are placed in arrays in the STORE where I can use them to spell out the
message line(s) the user requested, with plenty of alternate images.