All five text lines display as photos
Added .done() to the .getJSON(), but I'm not certain all of the problems
with async api calls are solved.

Rudimentary handling of spaces is done. All of the supported puctuation
is working.

Added function to handle getting the photo urls from the STORE and
adding them into the DOM.

Still to do: randomize the photos and cycle through them when clicked,
add the screen shot function, improve appearance especially
auto-resizing the photos for small screens and when there's too much
text on a single line. Need hard character limits on how much text can
be typed in line.