Stop errors of undefined URLs in the STORE
Added a test for undefined STORE.apiPhotos for characters whose API
calls have not happened yet. Before, I got an error of trying to get the
first URL in the array for every unique character in the text input
except for the last one to complete.

Also made the size of the photos a function of the viewport width, which
makes the letters display (if not well) on any size or shape of device
screen.

Set the spaces to half the width of the photos. Updated the Instructions
page to indicate that each line can only hold 14 characters max. I will
add a countdown counter, similar to Twitter, to let the user know how
much room is left.