Add Shuffle button to Ransom Note page
I noticed that clicking Back and then Continue without editing the text
caused all of the characters to select new random photos from the STORE
(without doing new API calls), and that the original app I'm recreating
did the same if you refresh the screen. It was a simple matter to add a
new button to do this directly, so I did, and updated the Instructions
to include the new button.