Final MVP
Greatly improved the CSS sugar.

Added counters, one for each line, to count down from 18 toward 0 as the
user types in text. Doesn't work if text is pasted in by clicking then
right-clicking and selecting Paste. Does work if text is pasted in with
Ctrl-v or Shift-Insert. Added a couple of lines explaining the counters.

Character photos are now properly centered vertically within the trays!

Added one last notice to the user, on the Ransom Note page: "Click any
photo to change it!" I noticed some users didn't get that it does that,
and that is the main point of the project.

Still would like to get the screen shot feature working. I have some
ideas now about how to get that wired up.