Getting 'pos' value now works in various browsers
Feedback from other Thinkful students showed that the photos were not
clickable in Firefox. Traced the problem to the assignment of the
charPos variable to targetCharacterPic[0].attributes[5].pos.value.
Firefox did not put the user-defined attribute in [5]. But it turns out
the [5] was not necessary; changing it to
targetCharacterPic[0].attributes.pos.value works fine in Chrome and in
Firefox.