Response to user feedback
The background image in the Main page did not load on one browser. I
changed the background color so that doesn't matter.

Added a box around the brief instructions and text fields in the Main
screen. Set the background color to off-white, 85% opacity.

Improved the instructions by moving the one paragraph of technical info
to a footnote, with smaller text. Edited the remaining instructions to
make it clearer that any text in any of the text fields is sufficient.