Better way to do the Distort feature
First pass at getting the Distort button to work on all pictures
retrieved worked, but required setting up 90 pairs of variables and 90
separate tags in the CSS file. My mentor showed me how to place the CSS
for this in the index.js file instead, using the .each() and .css()
methods. Also eliminated the need for separate numbered classes assigned
to each photo.